+++ webpack-rsc/src/lib/client-manifest.ts
